extract color-function
