+++ src/modules/olMap/services/StyleService.js
@@ -1,6 +1,7 @@
 import { getUid } from 'ol';
 import { $injector } from '../../../injection';
-import { markerStyleFunction, highlightStyleFunction, highlightTemporaryStyleFunction, measureStyleFunction, nullStyleFunction, lineStyleFunction, polygonStyleFunction, textStyleFunction, rgbToHex, markerScaleToKeyword, getStyleArray, geojsonStyleFunction, defaultStyleFunction } from '../utils/olStyleUtils';
+import { rgbToHex } from '../../../utils/colors';
+import { markerStyleFunction, highlightStyleFunction, highlightTemporaryStyleFunction, measureStyleFunction, nullStyleFunction, lineStyleFunction, polygonStyleFunction, textStyleFunction, markerScaleToKeyword, getStyleArray, geojsonStyleFunction, defaultStyleFunction } from '../utils/olStyleUtils';
